% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastgeojson.R
\name{disable_fast_json}
\alias{disable_fast_json}
\title{Disable Fast GeoJSON Serialization}
\usage{
disable_fast_json()
}
\value{
No return value. Called for the side effect of restoring the original
\code{jsonlite::toJSON} function.
}
\description{
Reverts the \code{jsonlite::toJSON} function to its original state.
}
